requirements.txt
# Django Core
Django==4.2.20
djangorestframework==3.15.2

# Database
mysqlclient==2.1.1
cryptography==44.0.1

# Cache
django-redis==5.4.0
redis==5.0.1

# Health checks
django-health-check==3.17.0

# Environment & Config
python-dotenv==1.0.1
gunicorn==23.0.0

# Development Tools (solo para desarrollo)
django-extensions==3.2.3
debugpy==1.8.5

# Utilities básicas
requests==2.32.4
Pillow==10.3.0
django-filter==23.5
python-dateutil==2.8.2

# Auditoría
django-simple-history==3.4.0

# AI/ML
openai==1.3.0
httpx==0.27.2

# WebSockets
channels==4.0.0
channels-redis==4.1.0

# Performance Profiling
django-silk==5.0.4

# System Monitoring
psutil==5.9.8

# Production Workers
gevent==25.9.1
greenlet==3.2.4
daphne==4.0.0
pymysql==1.1.0
structlog==23.2.0

# API Documentation
drf-spectacular==0.27.0
